Start Menu Script Bugfix
Fixed a bug that caused no ClickEvents being registereded after the UI GameObject was disabled and reenabled. This was caused by references to the UI Toolkit, that was initialized in Awake(), becoming Null when disabled and never getting reinitialized when the GameObject was reenabled. This was fixed by moving the initialization to OnEnable()