Some fixes
Readded audio to the pushable objects after it was removed.

Fixed an issue where the open book placeholder appeared in the ground by adding a yCorrection variable to the ChangeAfterFall script. Kept public as this issue might appear in other objects with this script and the value needed to fix could be different for those.