fix for being pulled on top of one-way colliders
Made a fix for the player being pulled on top of a one-way collider like the one used for the chair.

This is done by changing the players layer to a different one when moving upward during a jump, which makes it possible to stop strange collision interactions during jumps.